# Performance summary accounting for the case of LLM benchmarks