Updated `, get performance` output to work with LLM benchmarks. 
LLM Server benchmarks contain new latency metrics: Time To First Token (TTFT) and Time To Output Token (TPOT) that are now also collected.
The classic benchmarks output looks like before:
```
['INVALID', 'target_qps=541000', '99.00_percentile_latency=16.381 milliseconds', 'target_latency=15.000 milliseconds', 'latency_cutoff_ratio=1.09', 'Completed_samples_per_second=541169.81']
```
The LLM benchmarks output now looks like this:
```
['VALID', 'target_qps=6', '99.00_percentile_latency=26.794 seconds', '99.00_percentile_first_token_latency=0.181 seconds', 'ttft_latency=2.000 seconds', 'cutoff_ratio_ttft=0.09', '99.00_percentile_time_to_output_token=50.743 milliseconds', 'tpot_latency=0.200 seconds', 'cutoff_ratio_tpot=0.25', 'Completed_samples_per_second=5.98']
```